# fix bw for elementwise ops