Improves b/w for elementwise ops by using more work per thread.

Benchmarks are on M4 max

bench | Pre GB/s | Post GB/s
----- | ----- | ------
binary int8|241.379 | 465.679
binary float16|385.017 | 471.069
binary float32 | 465.569 | 475.750
unary int8 | 164.452| 460.587
unary float16| 292.264 | 465.375
unary float32 |444.194 | 466.945
copy uint8|  164.307 |460.884
copy int16|292.169 |464.882
copy int32 | 444.743 | 465.786
compile gelu float16 | 198.896 | 459.221
compile gelu float32 | 318.331 | 390.553